[Fix] #96 - 리뷰 반영
