+++ Wable-iOS/Infra/Network/OAuth/OAuthErrorMonitor.swift
@@ -27,22 +27,11 @@ final class OAuthErrorMonitor: EventMonitor {
     
     /// 서버로부터 응답을 받은 경우 호출되는 메서드
     func request<Value>(_ request: DataRequest, didParseResponse response: DataResponse<Value, AFError>) {
-        if let dataString = response.data?.toString,
+        
+        if let data = response.data,
+           let dataString = String.init(data: data, encoding: .utf8),
            response.error?.responseCode == 401 && dataString.contains(WableError.unauthorizedToken.rawValue) {
             isUnauthorized = true
         }
-    }
-}
-
-
-// MARK: - Extension
-
-extension Data {
-    var toString: String {
-        guard let object = try? JSONSerialization.jsonObject(with: self, options: []),
-              let data = try? JSONSerialization.data(withJSONObject: object, options: [.prettyPrinted]),
-              let dataString = NSString(data: data, encoding: String.Encoding.utf8.rawValue) else { return "" }
-        
-        return dataString as String
     }
 }
