+++ Wable-iOS/Infra/Network/OAuth/Repository/OAuthRepository.swift
@@ -1,0 +1,14 @@
+//
+//  OAuthRepository.swift
+//  Wable-iOS
+//
+//  Created by YOUJIM on 2/28/25.
+//
+
+
+import Combine
+import Foundation
+
+protocol OAuthRepository {
+    func updateTokenStatus() -> AnyPublisher<Token, WableError>
+}
