+++ Wable-iOS/Infra/Network/OAuth/Repository/OAuthRepositoryImpl.swift
@@ -1,0 +1,26 @@
+//
+//  OAuthRepositoryImpl.swift
+//  Wable-iOS
+//
+//  Created by YOUJIM on 2/28/25.
+//
+
+
+import Combine
+import Foundation
+
+import CombineMoya
+import Moya
+
+final class OAuthRepositoryImpl: OAuthRepository {
+    func updateTokenStatus() -> AnyPublisher<Token, WableError> {
+        return provider.request(
+            .fetchTokenStatus,
+            for: DTO.Response.UpdateToken.self
+        )
+        .map(LoginMapper.toDomain)
+        .mapWableError()
+    }
+    
+    private let provider = APIProvider<LoginTargetType>()
+}
