+++ Wable-iOS/Infra/Network/TargetType/ContentLikedTargetType.swift
@@ -16,7 +16,6 @@ enum ContentLikedTargetType {
 }
 
 extension ContentLikedTargetType: BaseTargetType {
-    
     var endPoint: String? {
         switch self {
         case .createContentLiked(contentID: let contentID):
